==> Code/01_data_extraction.ipynb <==
## This notebook will extract all "political" transcripts collected by Benjamin Litterer. It explored different labeling strategies as well. 
- Author: Bowen Yi

import os
os.environ["CUDA_DEVICE_ORDER"]="PCI_BUS_ID" 
device = 2
os.environ["CUDA_VISIBLE_DEVICES"]= str(device)
import torch 
import torch.nn as nn
device = torch.device("cuda:" + str(device) if torch.cuda.is_available() else "cpu")

import pandas as pd
import time
from multiprocessing import Pool
import os
import re
import numpy as np
import pycld2 as cld2  # language recognition 
from sklearn.model_selection import train_test_split
from sklearn.metrics import f1_score
from collections import Counter
# Let Jupyter Notebook show more content under cell
pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)
pd.set_option('display.width', None)
pd.set_option('display.max_colwidth', None)

import pandas as pd

## 1. Read in the raw unlabeled data

df = pd.read_json('/shared/3/projects/benlitterer/podcastData/processed/floydMonth/floydMonthData.jsonl', orient="records", lines=True)

df.head(1)

soted_dates = sorted(df['pubDate'])

df.columns

df.shape

## 2. Labeling strategy: 
### - A podcast is "politics" if it only has one single category assigned, and that being "politics". If none of its 10 category has "politics", it's non-politics
### - As we rely on weakly labeled data, this can keep our labeled data as clean as possible


def label_pol(row):
    if str(row['category1']).lower() == 'politics' and all(pd.isnull(row[col]) for col in ['category2', 'category3', 'category4','category5','category6','category7','category8','category9', 'category10']):
        return 1
    elif not any('politics' in str(category).lower() for category in row[['category1', 'category2', 'category3', 'category4', 'category5', 'category6', 'category7', 'category8', 'category9', 'category10']]):
        return 0
    else:
        return 2

pd.options.mode.chained_assignment = None
df['is_politics'] = df.apply(label_pol, axis=1)

df[df["is_politics"]==0].shape

df[df["is_politics"]==1].shape

## The first method yielded too limited positive samples. Thus, we label a podcast as politics as long as the top 1 category is "politics"

def label_pol_new(row):
    if str(row['category1']).lower() == 'politics':
        return 1
    elif not any('politics' in str(category).lower() for category in row[['category1', 'category2', 'category3', 'category4', 'category5', 'category6', 'category7', 'category8', 'category9', 'category10']]):
        return 0
    else:
        return 2

pd.options.mode.chained_assignment = None
df['is_politics_new'] = df.apply(label_pol_new, axis=1)

df[df["is_politics_new"]==1].shape

## Also too limited data. Try the third strategy. We label a podcast as politics if its top 3 category includes "politics"

def label_pol_third(row):
    if str(row['category1']).lower() == 'politics' or str(row['category2']).lower() == 'politics' or str(row['category3']).lower() == 'politics':
        return 1
    elif not any('politics' in str(category).lower() for category in row[['category1', 'category2', 'category3', 'category4', 'category5', 'category6', 'category7', 'category8', 'category9', 'category10']]):
        return 0
    else:
        return 2

pd.options.mode.chained_assignment = None
df['is_politics_top_3'] = df.apply(label_pol_third, axis=1)

df.columns

df[df["is_politics_top_3"]==1].shape

## By using this strategy, we have 7792 transcripts as "politics". That's enough data to start with

df_pol = df[df["is_politics_top_3"]==1].copy()

df_pol.to_csv("pol_trans.csv", index=False)

import pandas as pd
df = pd.read_csv("pol_trans.csv", usecols = ['transcript','enclosure','potentialOutPath','category1','category2', 'category3', 'category4', 'category5', 'category6', 'category7', 'category8', 'category9', 'category10'])


num_rows = df.shape[0]
chunk_size = num_rows // 7

dataframes = []
for i in range(0, num_rows, chunk_size):
    dataframes.append(df.iloc[i:i + chunk_size])

extra_rows = num_rows % 7
if extra_rows > 0:
    dataframes[-1] = pd.concat([dataframes[-1], df.iloc[-extra_rows:]])

for idx, dataframe in enumerate(dataframes):
    filename = f"pol_transcripts_{idx + 1}.csv"
    dataframe.to_csv(filename, index=False)
