==> Code/02_preprocessing.ipynb <==
## This notebook will preprocess transcripts such that they are easier to work with downstream models and annotate
- Author: Bowen Yi


import pandas as pd
import re
import pycld2 as cld2  # to detect foreign language in transcript
import cld3

df = pd.read_csv("pol_trans.csv", usecols = ['transcript','enclosure','potentialOutPath','category1','category2', 'category3', 'category4', 'category5', 'category6', 'category7', 'category8', 'category9', 'category10'])


df = df.dropna(subset=['transcript'])


df.shape


df = df.dropna(subset=['enclosure'])


df.shape


df = df.sample(frac=1, random_state=1)


def preprocess_transcript(text, words_to_remove=200, lower_bound_word_count=400):
    text = re.sub(r'\[.*?\]', '', text)
    text = re.sub(r'\(.*?\)', '', text)
    text = re.sub(r'\*.*?\*', '', text)
    text = text.replace('#', '')
    text = text.encode("ascii", "ignore").decode()
    
    if len(text.split()) <= lower_bound_word_count:
        return ""
    
    # This is for chunking transcripts later. We only chunk at the end of a sentence
    sentences = re.split(r'(?<=[.!?]) ', text)
    sentences = [sentence + ' ' for sentence in sentences]
    
    # When no sentences are found, or the text doesn't contain proper sentence-ending punctuation, ignore this transcript
    if len(sentences) <= 1:
        return ""
    
    # Remove the first and last ~200 words from the transcript
    start_index = 0
    word_count = 0
    for i, sentence in enumerate(sentences):
        word_count += len(sentence.split())
        if word_count >= words_to_remove:
            start_index = i
            break
    
    # Remove the last ~200 words from the transcript
    end_index = len(sentences) - 1
    word_count = 0
    for i in range(len(sentences) - 1, -1, -1):
        word_count += len(sentences[i].split())
        if word_count >= words_to_remove:
            end_index = i
            break
    
    if end_index <= start_index or start_index == 0 or end_index == len(sentences) - 1:
        return ""

    filtered_sentences = sentences[start_index + 1:end_index]
    return filtered_sentences



df['processed_text'] = df['transcript'].apply(preprocess_transcript)



# Remove short transcripts


def too_short(text, lower_bound=200):
    length = 0
    for sentence in text:
        length += len(sentence.split())
        if length > lower_bound:
            return False
    return True


df["too_short"] = df["processed_text"].apply(too_short)



df[df["too_short"]==True].shape


df = df[df["too_short"]==False]

df.shape

## Remove non-English transcripts

def is_eng(trans):
    text = ""
    for sent in trans:
        text += sent
    
    isReliable, textBytesFound, details, vectors = cld2.detect(text, returnVectors=True)
    if details[0][0] == "ENGLISH" and details[1][0] == "Unknown" and details[2][0] == "Unknown":
        return True
    else:
        return False
    

df["is_eng"] = df['processed_text'].apply(is_eng)



df = df[df["is_eng"]==1]


df = df.drop(['is_eng', 'transcript','too_short'], axis=1)


df.columns

df.processed_text[0]

## Convert processed_text into string

def convert_processed_to_str(text):
    res = ''
    for sentence in text:
        res += sentence
        
    return res
    

df['stringed_text'] = df['processed_text'].apply(convert_processed_to_str) 


df.columns

## Reorganize the dataframe columns

organized_data = {
    'transcript_to_chunk': df.processed_text,
    'path':df.potentialOutPath,
    'url':df.enclosure,
    'cat1':df.category1,
    'cat2':df.category2,
    'cat3':df.category3,
    'cat4':df.category4,
    'cat5':df.category5,
    'cat6':df.category6,
    'cat7':df.category7,
    'cat8':df.category8,
    'cat9':df.category9,
    'cat10':df.category10,
    'transcript_annotate':df.stringed_text,
#     'human_label':''
#     'note': ''
    
}

df_final = pd.DataFrame(organized_data)


df_final['human_label'] = ''
df_final['note'] = ''


df_final.head(1)

df_final = df_final.sample(frac=1/6, random_state=1).reset_index(drop=True)

df_final.to_csv('transcripts_to_annotate.csv',index=False)

df_final.shape

df_final.columns

df_final = df_final[df_final['transcript_annotate'].str.strip() != '']

df_final.shape
