==> Code/03_chunk_and_train.ipynb <==
## In this notebook, we attempted to finetune two large language models (miniLM, RoBERTa) to classify transcript chunks as right wing or left wing. 

- Author: Bowen Yi

import os
os.environ["CUDA_DEVICE_ORDER"]="PCI_BUS_ID" 
device = 1
os.environ["CUDA_VISIBLE_DEVICES"]= str(device)

import pandas as pd
import os
import re
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.metrics import f1_score
from collections import Counter

pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)
pd.set_option('display.width', None)
pd.set_option('display.max_colwidth', None)

df = pd.read_csv("07_Data_annotated_transcript.csv")

df.head(1)

df[df['Human Label']==1].shape

df[df['Human Label']==0].shape

df[df['transcript_to_chunk'].isna()].shape

df[~df['transcript_to_chunk'].isna()].shape

df = df[~df['transcript_to_chunk'].isna()]

df.shape

## 1. Split annotated transcripts to train, dev, and test sets (7:1:2)

df_train, df_dev_test = train_test_split(df, test_size=0.3, random_state=1)

df_dev, df_test = train_test_split(df_dev_test, test_size=2/3, random_state=1)

## 2. Chunk transcript

def chunk_transcript(text, chunk_size=100, max_size=256):
    chunks = []
    current_chunk = ''
    current_words_count = 0 

    for sentence in text:
        words_in_sentence = len(sentence.split())
        
        if current_words_count + words_in_sentence > chunk_size:
            chunks.append(current_chunk)
            current_chunk = sentence
            current_words_count = words_in_sentence
        else:
            current_chunk += sentence
            current_words_count += words_in_sentence

    if current_chunk.strip():
        if chunks and len(chunks[-1].split()) + len(current_chunk.split()) <= max_size:
               chunks[-1] += '' + current_chunk
        else:
               chunks.append(current_chunk.strip())

    return chunks



df.columns

### 2.1 Chunk train set

train_texts = []
train_labels = []
train_paths = []
train_chunk_ind = []
train_cat1 = []
train_cat2 = []
train_cat3 = []
train_cat4 = []
train_cat5 = []
train_cat6 = []
train_cat7 = []
train_cat8 = []
train_cat9 = []
train_cat10 = []

for _, row in df_train.iterrows():
    chunks = chunk_transcript(row['transcript_to_chunk'])
    chunk_count = len(chunks)
    
    if chunks:
        train_texts.extend(chunks)
        train_labels.extend([row["Human Label"]] * chunk_count)
        train_paths.extend([row["path"]] * chunk_count)

        
        chunk_ind = list(range(1, chunk_count+1))
        train_chunk_ind.extend(chunk_ind)

        train_cat1.extend([row["cat1"]] * chunk_count)
        train_cat2.extend([row["cat2"]] * chunk_count)
        train_cat3.extend([row["cat3"]] * chunk_count)
        train_cat4.extend([row["cat4"]] * chunk_count)
        train_cat5.extend([row["cat5"]] * chunk_count)
        train_cat6.extend([row["cat6"]] * chunk_count)
        train_cat7.extend([row["cat7"]] * chunk_count)
        train_cat8.extend([row["cat8"]] * chunk_count)
        train_cat9.extend([row["cat9"]] * chunk_count)
        train_cat10.extend([row["cat10"]] * chunk_count)
        

train_data = {
    'text':train_texts,
    'label':train_labels,
    'path':train_paths,
    'chunk_pos':train_chunk_ind,
    'cat1':train_cat1,
    'cat2':train_cat2,
    'cat3':train_cat3,
    'cat4':train_cat4,
    'cat5':train_cat5,
    'cat6':train_cat6,
    'cat7':train_cat7,
    'cat8':train_cat8,
    'cat9':train_cat9,
    'cat10':train_cat10,
}

df_chunks_train = pd.DataFrame(train_data)
df_chunks_train = df_chunks_train[df_chunks_train['text'].str.strip() != '']
df_chunks_train = df_chunks_train.reset_index(drop=True)
df_chunks_train = df_chunks_train.sample(frac=1, random_state=1).reset_index(drop=True)



df_chunks_train.shape

### 2.2 Chunk dev

dev_texts = []
dev_labels = []
dev_paths = []
dev_chunk_ind = []
dev_cat1 = []
dev_cat2 = []
dev_cat3 = []
dev_cat4 = []
dev_cat5 = []
dev_cat6 = []
dev_cat7 = []
dev_cat8 = []
dev_cat9 = []
dev_cat10 = []

for _, row in df_dev.iterrows():
    chunks = chunk_transcript(row['transcript_to_chunk'])
    chunk_count = len(chunks)
    
    if chunks:
        dev_texts.extend(chunks)
        dev_labels.extend([row["Human Label"]] * chunk_count)
        dev_paths.extend([row["path"]] * chunk_count)
        
        chunk_ind = list(range(1, chunk_count+1))
        dev_chunk_ind.extend(chunk_ind)

        dev_cat1.extend([row["cat1"]] * chunk_count)
        dev_cat2.extend([row["cat2"]] * chunk_count)
        dev_cat3.extend([row["cat3"]] * chunk_count)
        dev_cat4.extend([row["cat4"]] * chunk_count)
        dev_cat5.extend([row["cat5"]] * chunk_count)
        dev_cat6.extend([row["cat6"]] * chunk_count)
        dev_cat7.extend([row["cat7"]] * chunk_count)
        dev_cat8.extend([row["cat8"]] * chunk_count)
        dev_cat9.extend([row["cat9"]] * chunk_count)
        dev_cat10.extend([row["cat10"]] * chunk_count)


dev_data = {
    'text':dev_texts,
    'label':dev_labels,
    'path':dev_paths,
    'chunk_pos':dev_chunk_ind,
    'cat1':dev_cat1,
    'cat2':dev_cat2,
    'cat3':dev_cat3,
    'cat4':dev_cat4,
    'cat5':dev_cat5,
    'cat6':dev_cat6,
    'cat7':dev_cat7,
    'cat8':dev_cat8,
    'cat9':dev_cat9,
    'cat10':dev_cat10,
}

df_chunks_dev = pd.DataFrame(dev_data)
df_chunks_dev = df_chunks_dev[df_chunks_dev['text'].str.strip() != '']
df_chunks_dev = df_chunks_dev.reset_index(drop=True)
df_chunks_dev = df_chunks_dev.sample(frac=1, random_state=1).reset_index(drop=True)



df_chunks_dev.shape

### 2.3 Chunk test

test_texts = []
test_labels = []
test_paths = []
test_chunk_ind = []
test_cat1 = []
test_cat2 = []
test_cat3 = []
test_cat4 = []
test_cat5 = []
test_cat6 = []
test_cat7 = []
test_cat8 = []
test_cat9 = []
test_cat10 = []

for _, row in df_test.iterrows():
    chunks = chunk_transcript(row['transcript_to_chunk'])
    chunk_count = len(chunks)
    
    if chunks:
        test_texts.extend(chunks)
        test_labels.extend([row["Human Label"]] * chunk_count)
        test_paths.extend([row["path"]] * chunk_count)
        
        chunk_ind = list(range(1, chunk_count+1))
        test_chunk_ind.extend(chunk_ind)

        test_cat1.extend([row["cat1"]] * chunk_count)
        test_cat2.extend([row["cat2"]] * chunk_count)
        test_cat3.extend([row["cat3"]] * chunk_count)
        test_cat4.extend([row["cat4"]] * chunk_count)
        test_cat5.extend([row["cat5"]] * chunk_count)
        test_cat6.extend([row["cat6"]] * chunk_count)
        test_cat7.extend([row["cat7"]] * chunk_count)
        test_cat8.extend([row["cat8"]] * chunk_count)
        test_cat9.extend([row["cat9"]] * chunk_count)
        test_cat10.extend([row["cat10"]] * chunk_count)


test_data = {
    'text':test_texts,
    'label':test_labels,
    'path':test_paths,
    'chunk_pos':test_chunk_ind,
    'cat1':test_cat1,
    'cat2':test_cat2,
    'cat3':test_cat3,
    'cat4':test_cat4,
    'cat5':test_cat5,
    'cat6':test_cat6,
    'cat7':test_cat7,
    'cat8':test_cat8,
    'cat9':test_cat9,
    'cat10':test_cat10,
}

df_chunks_test = pd.DataFrame(test_data)
df_chunks_test = df_chunks_test[df_chunks_test['text'].str.strip() != '']
df_chunks_test = df_chunks_test.reset_index(drop=True)
df_chunks_test = df_chunks_test.sample(frac=1, random_state=1).reset_index(drop=True)



df_chunks_test.shape

df_chunks_test.to_csv("df_chunks_test.csv", index=False)
df_chunks_dev.to_csv("df_chunks_dev.csv", index=False)
df_chunks_train.to_csv('df_chunks_train.csv', index=False)

# df_chunks_train = pd.read_csv('df_chunks_train.csv')
# df_chunks_dev = pd.read_csv('df_chunks_dev.csv')
# df_chunks_test = pd.read_csv('df_chunks_test.csv')

# 3. Fine tune a model on labeled chunks

from datasets import Dataset
from transformers import AutoModelForSequenceClassification, AutoTokenizer
from transformers import TrainingArguments, Trainer, set_seed 
from sklearn.model_selection import train_test_split
from sklearn.metrics import confusion_matrix, precision_recall_fscore_support, accuracy_score


ds_train = Dataset.from_pandas(df_chunks_train[['text', 'label']])
ds_dev = Dataset.from_pandas(df_chunks_dev[['text', 'label']])
ds_test = Dataset.from_pandas(df_chunks_test[['text', 'label']])


tokenizer = AutoTokenizer.from_pretrained("microsoft/MiniLM-L12-H384-uncased")
def tokenize_function(examples):
    return tokenizer(examples["text"], max_length=512, padding="max_length", truncation=True)
ds_train = ds_train.map(tokenize_function, batched=True, batch_size=512)
ds_dev = ds_dev.map(tokenize_function, batched=True, batch_size=512)
ds_test = ds_test.map(tokenize_function, batched=True, batch_size=512)


from transformers.integrations import WandbCallback
os.environ["WANDB_PROJECT"]="miniLM_politics"
os.environ["WANDB_LOG_MODEL"] = "checkpoint"

model = AutoModelForSequenceClassification.from_pretrained("microsoft/MiniLM-L12-H384-uncased", num_labels=2)

output_dir = "/shared/3/projects/bowenyi/where-is-news/595_486_Final_Project/model_output_miniLM"   
seed = 1

def compute_metrics(eval_pred):
    logits, labels = eval_pred
    predictions = np.argmax(logits, axis=-1)
    
    conf_matrix = confusion_matrix(y_true=labels, y_pred=predictions)
    TN, FP, FN, TP = conf_matrix.ravel()
    
    TPR = TP / (TP + FN) if (TP + FN) > 0 else 0.0
    FPR = FP / (FP + TN) if (FP + TN) > 0 else 0.0
    TNR = TN / (TN + FP) if (TN + FP) > 0 else 0.0
    FNR = FN / (TP + FN) if (TP + FN) > 0 else 0.0

    precision, recall, f1, _ = precision_recall_fscore_support(labels, predictions, average='binary')
    accuracy = accuracy_score(labels, predictions)
    
    return {
        'accuracy': accuracy,
        'f1': f1,
        'precision': precision,
        'recall': recall,
        'true_negative_rate': TNR,
        'false_positive_rate': FPR,
        'false_negative_rate': FNR,
        'true_positive_rate': TPR
    }


training_args = TrainingArguments(
    output_dir=output_dir,
    overwrite_output_dir=True,
    learning_rate=2e-5,
    per_device_train_batch_size=32,
    per_device_eval_batch_size=64,
    num_train_epochs=8,  
    weight_decay=0.01,
    do_train=True,
    do_eval=True,
    seed=seed,
    save_strategy='steps',
    save_steps=500,
    evaluation_strategy='steps',
    eval_steps=500,
    logging_dir=output_dir + 'logs/',
    logging_strategy='steps',
    logging_steps=500,
    load_best_model_at_end=True,
    metric_for_best_model="f1",
    run_name='podcasts-study' + str(seed),
    report_to="wandb"
)


trainer = Trainer(
    model=model,    
    args=training_args,
    train_dataset=ds_train,
    eval_dataset=ds_dev,    
    tokenizer=tokenizer,
    compute_metrics=compute_metrics
)


trainer.add_callback(WandbCallback())
trainer.train()


test_result = trainer.evaluate(ds_test)
print(test_result)
wandb.finish()


### We noticed that training model on transcript chunk isn't good enough. The highest F-1 score is not even better than chance, which indicates a weak training signal. Considering that we annotate instances at transcript level but train and evaluate at chunk level, the model performance makes sense. 

## 4. Train miniLM with Reddit data and transcript chunks
- Reddit data are better labeled than transcript chunks. We hope this can make our training signal stronger. 
- Reddit data source: https://www.kaggle.com/datasets/neelgajare/liberals-vs-conservatives-on-reddit-13000-posts

df_red = pd.read_csv('08_reddit.csv')

df_red.shape

df_red.head(1)

### This is the best external dataset we can find to augment our weakly labeled dataset. We treat Liberal as left wing (0), and Conservative as right wing (1). Although Liberal isn't equal to leftist, we manually examined the dataset and found their definition is pretty close (in this dataset). 

df_red.columns

red_text = df_red.Title.to_list()

red_label = df_red['Political Lean'].to_list()

red_labels = [1 if lean == 'Conservative' else 0 for lean in red_label]

del df_red

df_red = pd.DataFrame({'text':red_text, 'label':red_labels})

df_red.head(1)

df_red.shape

## To further balance the impact of weakly labeled chunks, we downsample the dataset of chunks such that it has around the same amount as Reddit Data. Also, we downsampled the number of left-wing chunks to make the class more balanced 

df_chunks = pd.concat([df_chunks_train, df_chunks_dev, df_chunks_test])


df_chunks.shape

df_chunks = df_chunks.sample(frac=0.15, random_state=1).reset_index(drop=True)

df_chunks.shape

df_chunks = pd.concat([df_chunks, df_red])

df_chunks[df_chunks['label']==1].shape

df_chunks[df_chunks['label']==0].shape

df_left = df_chunks[df_chunks['label']==0].sample(n=16000, random_state=1).reset_index(drop=True)


df_left.shape

df_right = df_chunks[df_chunks['label']==1].reset_index(drop=True)


df_right.shape

df_chunks = pd.concat([df_left, df_right])

df_chunks.shape

df_train, df_dev_test = train_test_split(df_chunks, test_size=0.3, random_state=1)

df_dev, df_test = train_test_split(df_dev_test, test_size=2/3, random_state=1)

## 4.1 Finetune a model on downsampled labeled chunks and all Reddit data
- We tried a more powerful model: RoBERTa-base

from transformers import RobertaTokenizerFast, RobertaForSequenceClassification, AutoConfig


ds_train = Dataset.from_pandas(df_train[['text', 'label']])
ds_dev = Dataset.from_pandas(df_dev[['text', 'label']])
ds_test = Dataset.from_pandas(df_test[['text', 'label']])


model_id = "roberta-base"
tokenizer = RobertaTokenizerFast.from_pretrained(model_id)

def tokenize_function(examples):
    return tokenizer(examples["text"], max_length=256, padding="max_length", truncation=True)
ds_train = ds_train.map(tokenize_function, batched=True, batch_size=512)
ds_dev = ds_dev.map(tokenize_function, batched=True, batch_size=512)
ds_test = ds_test.map(tokenize_function, batched=True, batch_size=512)

model = RobertaForSequenceClassification.from_pretrained(model_id, num_labels=2)

import wandb
wandb.login()
from transformers.integrations import WandbCallback
os.environ["WANDB_PROJECT"]="roberta_base_politics"
os.environ["WANDB_LOG_MODEL"] = "checkpoint"

output_dir = "/shared/3/projects/bowenyi/where-is-news/595_486_Final_Project/model_output_roberta_base"   
seed = 1


def compute_metrics(eval_pred):
    logits, labels = eval_pred
    predictions = np.argmax(logits, axis=-1)
    
    conf_matrix = confusion_matrix(y_true=labels, y_pred=predictions)
    TN, FP, FN, TP = conf_matrix.ravel()
    
    TPR = TP / (TP + FN) if (TP + FN) > 0 else 0.0
    FPR = FP / (FP + TN) if (FP + TN) > 0 else 0.0
    TNR = TN / (TN + FP) if (TN + FP) > 0 else 0.0
    FNR = FN / (TP + FN) if (TP + FN) > 0 else 0.0

    precision, recall, f1, _ = precision_recall_fscore_support(labels, predictions, average='binary')
    accuracy = accuracy_score(labels, predictions)
    
    return {
        'accuracy': accuracy,
        'f1': f1,
        'precision': precision,
        'recall': recall,
        'true_negative_rate': TNR,
        'false_positive_rate': FPR,
        'false_negative_rate': FNR,
        'true_positive_rate': TPR
    }

training_args = TrainingArguments(
    output_dir=output_dir,
    overwrite_output_dir=True,
    learning_rate=2e-5,
    per_device_train_batch_size=64,
    per_device_eval_batch_size=64,
    num_train_epochs=3,
    weight_decay=0.01,
    do_train=True,
    do_eval=True,
    seed=seed,
    save_strategy='steps',
    save_steps=100,
    evaluation_strategy='steps',
    eval_steps=100,
    logging_dir=output_dir + 'logs/',
    logging_strategy='steps',
    logging_steps=100,
    load_best_model_at_end=True,
    metric_for_best_model="f1",
    run_name='podcasts-study' + str(seed),
    report_to="wandb"
)

trainer = Trainer(
    model=model,    
    args=training_args,
    train_dataset=ds_train,
    eval_dataset=ds_dev,    
    tokenizer=tokenizer,
    compute_metrics=compute_metrics
)


trainer.add_callback(WandbCallback())
trainer.train()
trainer.evaluate(ds_test)
wandb.finish()

## This strategy achieved a better F-1 score. Our next step is to manually check 1000 samples predicted by RoBERTa. Then, we can set a proper decision threshold for right wing and left wing, instead of using the default 0.5. 

## We should have used a separate set of data for calibration. But due to high-quality data scarcity, we used the test set to calibrate

## We divided the entire test set to 10 ten-percent bins. Then, we randomly sampled 100 instanced in each 10 percent bin for manually checking

## 4.2 Produce file for calibration

import torch
from scipy.special import softmax 

predictions = trainer.predict(ds_test)

probs = softmax(predictions.predictions, axis=1)[:, 1]
df_test['prob'] = probs


df_test['stratum'] = pd.cut(df_test['prob'], bins=[0, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0], 
                       labels=['0-10%', '10-20%', '20-30%', '30-40%', '40-50%', '50-60%', '60-70%', '70-80%', '80-90%', '90-100%'], 
                       include_lowest=True)

df_test['ground_truth'] = ''
df_test['note'] = ''
df_test_cp = df_test.copy()

annotate_df = df_test_cp.groupby('stratum').apply(lambda x: x.sample(n=100, random_state=1)).reset_index(drop=True)



annotate_df.to_csv('09_RoBERTa_annotate.csv')


## After manually checking, we decide 0.7 would be a decision threshold. If an input text receives score above 0.7, it's right-wing. If lower, it's left-wing. 

## 4.3 Out of curiosity, we tested a finetuned BERT-based model specifically for detecting political leaning. The model is introduced by an EMNLP article in 2020.

pol_tokenizer = AutoTokenizer.from_pretrained("bert-base-cased")

pol_bert = AutoModelForSequenceClassification.from_pretrained("bucketresearch/politicalBiasBERT")


texts = [entry['text'] for entry in ds_test]  
tokenized_texts = tokenizer(texts, max_length=256, padding="max_length", truncation=True)

tokenized_texts_dataset = Dataset.from_dict(tokenized_texts)

polbert_trainer = Trainer(model=model)

polbert_predictions = polbert_trainer.predict(tokenized_texts_dataset)




from sklearn.metrics import f1_score, accuracy_score


polbert_probs = torch.nn.functional.softmax(torch.tensor(polbert_predictions.predictions), dim=-1).numpy()

predicted_labels = np.argmax(polbert_probs, axis=1)

true_labels = np.array([entry['label'] for entry in ds_test])  # Ensure the key 'label' matches the key in ds_test

f1 = f1_score(true_labels, predicted_labels)
accuracy = accuracy_score(true_labels, predicted_labels)

print(f"F1 Score: {f1}")
print(f"Accuracy: {accuracy}")

## Seems that the pretrained model didn't perform better than our own finetuned model. (Also to show effort), we'll use our own model

trainer.save_model(output_dir = "best_roberta")
