==> Code/06_tfidf.ipynb <==
import pandas as pd
import numpy as np
from scipy.special import softmax
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import confusion_matrix
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics import f1_score
from sklearn.model_selection import train_test_split

pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)
pd.set_option('display.width', None)
pd.set_option('display.max_colwidth', None)

df_red = pd.read_csv('08_reddit.csv')
df_chunks_train = pd.read_csv('df_chunks_train.csv')
df_chunks_dev = pd.read_csv('df_chunks_dev.csv')
df_chunks_test = pd.read_csv('df_chunks_test.csv')


df_chunks_train.head(1)

df_chunks = pd.concat([df_chunks_train, df_chunks_dev, df_chunks_test])


df_chunks = df_chunks.sample(frac=0.15, random_state=1).reset_index(drop=True)


df_chunks.shape

df_chunks = pd.concat([df_chunks, df_red])

df_chunks.shape

df_left = df_chunks[df_chunks['label']==0].sample(n=16000, random_state=1).reset_index(drop=True)


df_right = df_chunks[df_chunks['label']==1].reset_index(drop=True)


df_chunks = pd.concat([df_left, df_right])

df_chunks.shape

from sklearn.model_selection import train_test_split

df_train, df_test = train_test_split(df_chunks, test_size=0.3, random_state=1)

X_train = list(df_train['text'])
y_train = list(df_train['label'])

X_test = list(df_test['text'])
y_test = list(df_test['label'])

vectorizer = TfidfVectorizer(ngram_range=(1,2), min_df=5)


X_train_vectorized = vectorizer.fit_transform(X_train)
X_test_vectorized = vectorizer.transform(X_test)

clf = LogisticRegression(max_iter=500, class_weight = "balanced", random_state=1, C=3.0) #, C=2.0


clf.fit(X_train_vectorized, y_train)


y_pred_train = clf.predict(X_train_vectorized)
print("F1 score on the train set:", f1_score(y_train, y_pred_train, average='binary'))

y_pred_test = clf.predict(X_test_vectorized)
print("F1 score on the test set:", f1_score(y_test, y_pred_test, average='binary'))

tn, fp, fn, tp = confusion_matrix(y_test, y_pred_test).ravel()

tpr = tp / (tp + fn)  # True Positive Rate (Recall)
fpr = fp / (fp + tn)  # False Positive Rate
tnr = tn / (tn + fp)  # True Negative Rate (Specificity)
fnr = fn / (fn + tp)  # False Negative Rate

print(f"True Positives: {tp}")
print(f"False Positives: {fp}")
print(f"True Negatives: {tn}")
print(f"False Negatives: {fn}")
print(f"True Positive Rate: {tpr:.4f}")
print(f"False Positive Rate: {fpr:.4f}")
print(f"True Negative Rate: {tnr:.4f}")
print(f"False Negative Rate: {fnr:.4f}")

term_counts = np.squeeze(np.asarray(X_test_vectorized.sum(axis=0)))

coefficients = clf.coef_[0]
impact_scores = coefficients * np.log(term_counts + 1)

feature_names = vectorizer.get_feature_names_out()

impact_scores_series = pd.Series(impact_scores, index=feature_names)

highest_impact = impact_scores_series.sort_values(ascending=False).head(100)
lowest_impact = impact_scores_series.sort_values(ascending=False).tail(100)

print("30 Highest Impact Terms:\n", highest_impact)
print("\n25 Lowest Impact Terms:\n", lowest_impact)
